Move examples in separate pages
