+++ packages/www/pages/examples/DateInput.tsx
@@ -1,0 +1,20 @@
+import { useEffect, useState } from "react";
+import { Example } from "../../components/Example";
+import { DateInput } from "../../components/examples/DateInput";
+
+export default function ExamplesPage() {
+  const [client, setClient] = useState(false);
+
+  useEffect(() => {
+    setClient(true);
+  }, []);
+
+  return (
+    <Example
+      file="/packages/www/components/examples/DateInput.tsx"
+      title="DateInput"
+    >
+      {client ? <DateInput /> : null}
+    </Example>
+  );
+}
