+++ packages/www/pages/examples/DateTimePicker.tsx
@@ -1,0 +1,20 @@
+import { useEffect, useState } from "react";
+import { Example } from "../../components/Example";
+import { DateTimePicker } from "../../components/examples/DateTimePicker";
+
+export default function ExamplesPage() {
+  const [client, setClient] = useState(false);
+
+  useEffect(() => {
+    setClient(true);
+  }, []);
+
+  return (
+    <Example
+      file="/packages/www/components/examples/DateTimePicker.tsx"
+      title="DateTimePicker"
+    >
+      {client ? <DateTimePicker /> : null}
+    </Example>
+  );
+}
